Display errors on earlier arguments first in signature selection
Summary:
When multiple arguments match to the same param, we were reliably revealing the errors against the later arguments first.

Make sure the order prioritizes earlier errors, but also add in an importance factor so we can elevate certain types of errors over other (and have order be the second-tier sort).

We still are getting the order wrong in the case of parameter ordering, since we use a map to store the param-arg relationships. Fix to that will be stacked on top of this. (Motivation for all of this is fixing kwargs and args type errors  - see task attached :P)

Reviewed By: dark

Differential Revision: D13578205

fbshipit-source-id: ea63125a20e0a5b8fa6dfbc40149791fe35b091e