Track given annotation in missing errors
Summary: It's worthwhile to keep the given annotation we threw a missing annotation error against, because we throw errors when given annotations are "Any" or if any piece of them are "Any". Rather than tracking the bool, keep the whole annotation so we can error properly when we see something like `Dict[Any, str]`.

Reviewed By: dark

Differential Revision: D13605369

fbshipit-source-id: 7937e7fc8f6b387cb6f6915a8abad2e7aa1edd0f