Compare sanitized accesses in errors
Summary:
Partially a RFC... not sure if there are any implications I haven't thought of.

This is a no-op for most accesses, but we do run into situations like:

```
def foo(x):  # error thrown against `x` here has access `$parameter$x`
   x = 1  # error thrown against `x` here has access `x`
```

This bit us when throwing against parameter errors in particular, but not sure if there is ever a case where we want to be comparing unsanitized/sanitized accesses as different.

Reviewed By: sinancepel

Differential Revision: D13615696

fbshipit-source-id: c726a54ac7bc9d92b6c22890fa1229edeb6565d3