Fix incorrect weakening caused by contradicting assertions
Summary:
While forwarding the state through the assertion of a ternary (or,
truly, any assertion in any statement that branches and the re-joins,
like `If`), the refined annotation was always accepted, even if it is
in contrast with the previous state. This means, for example, that an
optional narrowed to a non-optional, because of a previous test, would
revert being an optional after the ternary:

```
 if x is not None:
   # x is not optional
   ... = x if x is None else ...
   # x is back to be an optional
```

Rectify the behaviour by having the assert not add to the resolution,
unless the new annotation is less or equal than the already
established one.

Also, throw away the new resolution at the end of a ternary: any
side-effects are local to the ternary and should not escape its scope.

Reviewed By: dkgi

Differential Revision: D13578618

fbshipit-source-id: fedb0dbcdb5ffb716a28ba97a8845819d9f4ce6a