Introduce opamcake archive to speed up and stabilize building binaries
Summary:
Almost all of our issues with flakiness on our CI come from failures in downloading and compiling our libraries (and opam itself).  We can avoid this, and the massive time penalty this imposes by building that once and then just downloading and using that state.  The trickiness that stopped this the first time I tried to do this was that there is actually a lot of times in the opam build that the position of other opam resources is stored absolutely (as opposed to relative to the root).  This is arguably a bug in opam as the root is supposed to be the point of reference, but in this case we can avoid it by doing the compiling in the same location as it will be used, a location that exists on all relevant platforms, a tmp directory.

I called this an opamcake because its a nice flattened version of the complex state of opam.

Reviewed By: dark

Differential Revision: D13634316

fbshipit-source-id: 879780b7307950f7886f726b8cacfdf0b000a2c3