Resolve dictionary literals
Summary:
We were erroring on (see also - attached task):

```
a.py

my_dict = {"a": 1}
```

```
b.py

from a import my_dict

def foo() -> typing.Dict[str, int]:
   return my_dict
```

This was simply because we did not recognize `{"a": 1}` when building the environment and therefore registered it as Top (unknown).

Resolving dict literals will allow these to get registered to globals as a dictionary rather than unknown, solving this issue on the surface (at least, making it consistent with how we treat all other unannotated globals).

------
------

Marginally related to this diff, but something this brought up:

The question of whether the way we deal with all unannotated globals is sound is still up for debate.

```a.py

x = 1
```

```
b.py

from a import x

def foo() -> int:
   return x

def bar() -> str:
   global x
   x = "string"
   return x
```

This will throw:
``Missing global annotation [5]: Globally accessible variable `export.x` has type `str` but no type is specified.`` on the global, but remain silent on both of the functions returning the global.

When the unannotated global assignment is not to a literal, we can't resolve the value when building the environment and default that global to be unknown (so we only ever have globals with a type but an unknown annotation when we encounter literals).

Some ways to approach this problem, since this is clearly not type safe:
1. Assume globals assigning to literals were also annotated as having the type of that literal. That would thrown an error when we try to assign something that isn't a subtype of the initial literal.
2. Only compare against the "original" annotation of globals when returning errors, so unannotated globals will be treated as unknown, in the same way as non-literal globals.

Reviewed By: dkgi

Differential Revision: D13643717

fbshipit-source-id: 542a31eaa010f36c6d6ddbe1f1cf68c0e1164a3c