Add "long" description to JSON error output
Summary:
It is sometime useful to display "long" error descriptions, and the
JSON output does not include such a field. Currently, Pyre determines
server-side whether to print a short or a detailed message, making
error message selection at the client difficult.

This diff adds an extra field to the returned JSON, that always
includes the detailed error description. This way, the error consumer
can decide whether to print a short or a detailed message.

Reviewed By: shannonzhu

Differential Revision: D13701485

fbshipit-source-id: 0e1968c622d7bf8e27f4d63bc39b4f28825bab2b