Delete outdated comment
Summary:
context: this was added when hotfixing the issue with paths including commas were passed in a comma-separated list.

Now that the hotfix flag `filter-directories-semicolon` has been renamed to `filter-directories` and the old comma separated one has been deleted, this comment is just confusing. Callers to this flag should not go away.

Reviewed By: dark

Differential Revision: D13704977

fbshipit-source-id: f62a8dde4047a3bb12b892db364e1702c916d6e3