Change behaviour of variant types against typing.Any
Summary:
A list of ints is not a list of Anys.

In other words, it is an error to provide a container of concrete
types where a container of typing.Any is expected: this is because the
caller might mutate the container and still respect `typing.Any`, but
now the original invariant would be violated.

This error will be raised in strict mode in most cases, but several
cases will slip through `Type.mismatch_with_any`. This is because that
function suppresses errors only if the actual and expected types are
the same parametric type: it does not account for subtypes.

Reviewed By: shannonzhu

Differential Revision: D13677820

fbshipit-source-id: ddb4bc18c4b27caa7a53bd07176ae54a8f5dfc1f