Fix error suppression for missing errors with unknown
Summary:
We care about suppressing if the given annotation is unknown (clearly something went wrong). If our inferred annotation is unknown, that is totally fine -- the inferred annotation is only shown in the messaging if it is something useful, anyway.

**Not knowing what annotation to suggest does not mean that the annotation is not missing in the first place.** We should still throw a missing annotation error!

Ran into these when fixing up analysis crashes (a cause of unknown given annotations).

Reviewed By: dkgi

Differential Revision: D13725756

fbshipit-source-id: 3a2744f30828a37dedd4af645fc4f4cafdcfa7e5