refactor AnalysisError.join to unify Top case
Summary:
We rely on the catch-all Top case for debugging purposes - namely, having Top only be created in that location ensures that
when we enable the debug logging section, we get detailed diagnostics on what went wrong exactly. The ad-hoc Top generation didn't have this,
so eliminate those cases.

The refactor was very straightforward in most cases, only leading to some duplicated pattern matching in one case.

Reviewed By: shannonzhu

Differential Revision: D13768978

fbshipit-source-id: 71b41cd6740ccb36e6dd3662bfcb0cafeeffd4e1