Do not propagate "Undeclared" across assignments
Summary:
When a variable whose type is "Undeclared", or partially "Undeclared",
is assigned to another one, the latter should not be considered
"Undeclared". After all, we just assigned to it, so it is definitely
declared.

This avoids propagating "Undeclared" errors forever; i.e., in the
following code, we should only warn on the first line, as that's the
only one that can be acted on. Erroring on the other lines would only
add to the noise.

```
a = something_undeclared
b = a
c = b
d = c
```

Reviewed By: shannonzhu

Differential Revision: D13708793

fbshipit-source-id: f7f5abd9b8042559f3066cb97428873ec45dc9fc