Skip config in fixme-all if no version present
Summary:
We don't care about fixmeing a local configuration unless there's a version # to remove and clean up after.

Wasn't sure if this was worth turning into a flag, but happy to do that if that's worthwhile.

Came up because I was trying to use this script to clean up the last of the local configurations in fbcode that still have a version override set (some from this recent upgrade, some from upgrades a while ago that we missed). Didn't want to build and run and fixme all these unnecessary local configurations that are already typechecking against the master fbcode version.

Reviewed By: sinancepel

Differential Revision: D13829654

fbshipit-source-id: 8cecb91557edc781188069cec36b4b99bbc6a690