Prevent truncation or loss of exception messages
Summary:
When the Pyre binary exists with an exception, it prints its status in
a format that clearly will not follow our logger format.

The previous code would fail to display these messages if a buffer had
been flushed, or contained other information, preventing the client
from displaying the line and essentially losing the message.

Change the logic in the client so that lines that do not match our
format are passed through without interpretation. This is the safest
approach to prevent loss of any messages that we do not expect.

Reviewed By: sinancepel, shannonzhu

Differential Revision: D13832397

fbshipit-source-id: f4b42d3ff0b84e953be81b91bee6dc35735fc80e