Add query to dump file dependencies to `.dot` file
Summary:
It is useful to dump file dependencies, either for debugging reasons,
or to understand large codebases.

This diff introduces a new query ("dump_dependencies") that, given a
filename, dumps the depenency graph of that module to a file called
`dependencies.dot`. The file can be post-processed with `graphviz` to
obtain a graphical representation of the dependencies.

Reviewed By: sinancepel

Differential Revision: D13797568

fbshipit-source-id: 511486891226e524aea5f7375c2340ff184ae15d