Simplify handling of deferred server requests
Summary:
Deferred server requests are always typecheck requests, so simplify
the state by having a list of deferred files to analyze.

Fully determine the list of transitive dependencies when a handle is
received for analysis, rather than as we go. To simply state
management, a File.Set.t is added to the state to avoid scanning the
list of dependencies repeatedly, as it grows larger.

Finally, abstract this behaviour behind a separate module to make sure
client/command code does not rely on internal state behaviour.

Reviewed By: sinancepel

Differential Revision: D13809004

fbshipit-source-id: 0236ac683be8a3e47779edecb523e2e42ca4f2b5