Fix pyre server's LSP parsing with extraneous header lines
Summary: The current version of LSP supports a `Content-Type` header, and in the future more header fields could be introduced. Currently, we just assume the line after `Content-Length` is an empty line, which breaks with new headers. This change allows us to consume an arbitrary number of header lines until we actually read an empty line.

Reviewed By: dark

Differential Revision: D13867927

fbshipit-source-id: 9d0ccba0ed1b096ee5fafdc9121ebd7871349545