Assume explicit annotation given for global literals
Summary:
This came up during the distillery sprint - there were cases of files full of constants that were throwing errors on every single global assignment. These would have to be codemodded to stick a `: int` in each line, for example, and this was irritating.

This change treats any globals that are assigned to literal values as having been explicitly annotated.

`x = 1` at toplevel will no longer error that `x` is a global and needs an annotation; it is treated the same as `x: int = 1`.

Any subsequent assignment to `x` that is not an int will error (so the first non-explicit assignment is the only one treated as explicit, to avoid crazy erroring later on).

(note to self: also mentioned in summary of D13643717)

Reviewed By: sinancepel

Differential Revision: D13821360

fbshipit-source-id: e8f3bfe6bc9f8f42914bf16745e6b670a2d193b2