Implement resolve literal for more expressions
Summary: Make this a bit more robust, now that what we consider resolvable literals is a bit more significant (see rest of stack). We also have had confusion in the past when seeing something like `x = {1: 2}` annotated to be unknown since we didn't support resolving this as a literal. Avoid this by supporting the obvious (and safe) cases.

Reviewed By: sinancepel

Differential Revision: D13822554

fbshipit-source-id: 370e7c81931491b2a1d3ea82eaf5696e969a798b