Stop suppressing sig select errors when return is not bound
Summary:
When supplying too few errors to a signature with type variables, the return type will not be bound and wind up getting suppressed.

We should only check whether the return annotation is resolved if the signature was found; otherwise we need to keep the missing annotation (or other) errors around.

Reviewed By: sinancepel

Differential Revision: D13880817

fbshipit-source-id: 53f1bbc0e4ed2f955b231fa8bfeb3e71f99c48d1