Support strings as request IDs in pyre server
Summary: The JSON-RPC spec (as of 2.0) permits the `id` field of a request to be [either an integer or string](https://www.jsonrpc.org/specification#request_object). Prior to this diff we only supported integer IDs, and requests with string IDs (such as those provided by pyls) would never receive a response.

Reviewed By: shannonzhu

Differential Revision: D13881268

fbshipit-source-id: 89b2f033c17a72fd05c6d32ef58863de25b7bf14