Accept hash as argument to fixme-all
Summary:
Upgrading via sandcastle will be equivalent to fixme-all, which should take care of the entire process given a particular hash.

Allow a hash to be optionally passed in to the fixme-all command. If `--sandcastle` is not set (to be implemented above this), do the upgrade locally with the hash.

Reviewed By: sinancepel

Differential Revision: D13904216

fbshipit-source-id: 72c1abc8b76ceaca674757dfa01d115a7cd9c3e9