Fix a NULL-pointer dereference bug in sqlite3 code
Summary:
Fix a NULL-pointer dereference bug in the sqlite3 code.

The `addr` field in a bucket can be NULL, even if the `hash` is
non-zero, if the bucked has been allocated to a key that has since
been deleted. In terms of `hh_hash_used_slots`, this is a "nonempty",
but also not "filled" slot.

Reviewed By: sinancepel

Differential Revision: D13889245

fbshipit-source-id: acf4225103c45e55879b6352f8504af6f6abe1b0