Restrict heuristic for inferring immutable constructor attrs
Summary:
pleasekillme2

Ran into this issue in D13908309 when I wrote:

```
class Foo:
   def __init__(self, x: int) ->
      if something:
         self.attribute = None
      else:
         self.attribute = x

```

And I got the error:

Incompatible attribute type: Attribute `attribute` of class `Foo` has type int but is used as type None

This was bizarre because not only was the `self.attribute = x` defined later on than the line triggering the incompatible attribute error, but it was also not annotated.

Turns out forward and access fold are all fine, but when we retrieve the attributes from a class we're special-casing values that are also parameters and treating them as immutable for the following situation:

```
class Foo:
   def __init__(self, x: int, y: int, z: int):
       self.x = x
       self.y = y
       self.z = z
```

So these wouldn't have to each be re-annotated.

Fixed this up so we only assume immutability if the attribute and parameter names match, and the assignemnt is done at the toplevel of the function (ie., not inside if/else branches)

Reviewed By: sinancepel

Differential Revision: D13928554

fbshipit-source-id: ea5d96369d1337fdd6f59a18463ca76b9d53afd8