Parse annotation should not return untracked annotations by default
Summary:
If the expression you are trying to parse into a type is not tracked, you should get `unknown` back from `parse_annotation` unless you know what you're doing and explicitly want to allow untracked annotations to be returned.

This systematically gets rid of untracked types floating around in analysis and causing crashes somewhere downstream (working on getting rid of cases where the flags are set to true without an obvious semantic reason... commented on the places where we definitely do want the flag set to true - the rest need investigation).

The only time we actually want the untracked annotation to be returned during typechecking is when we are checking the literal expression itself (ie., the define where a signature annotation is wrong, etc) in which case we call `parse_and_check` to add the undefined type error to state. This is already happening in most cases.

Reviewed By: dkgi

Differential Revision: D13946079

fbshipit-source-id: 54e135c9b231c7484f9dc37e3b7d81316ac7a6e1