Dramatically reduce time needed to purge keys
Summary:
When receiving a request to purge several thousands files, the
previous code would take several minutes to complete. By deduplicating
the access list, the time required drops to a few seconds only.

This is a stopgap measure until the lists in shared memory can be
fully audited for performance.

Reviewed By: sinancepel

Differential Revision: D13965574

fbshipit-source-id: c3183f9d5baecee8f21d9f32a86077d2b490af5d