Print unadorned errors to stderr
Summary:
D13832397 added code to pass-through messages printed by the binary
that do not match the expected format.

That diff started printing these errors on stdout though, rather than
on stderr. That causes problems because proper results are possibly
interspersed between extraneous messages.

This diff changes that behaviour so errors are properly printed to
stderr. Note that using stderr directly does not work, because of the
tricks we do with the logging infrastructure: hence the logger is used
instead.

Reviewed By: shannonzhu

Differential Revision: D13975706

fbshipit-source-id: b82e5d8c5484422bda80bd4651a094b0fa4ce8a3