Make dependency lists sets wherever possible
Summary:
As part of the server consistency testing, it was discovered that a
tangible number of discrepancies in the shared memory layout is due to
the dependency lists diverging.

In addition to that, dependency lists are not tested for duplicates.

Migrating all such lists to sets (and trees where serialization is
required) fixes duplication problems and makes the shared memory
easier to compare.

Reviewed By: sinancepel

Differential Revision: D13984583

fbshipit-source-id: 7384f5e7bd3e9c890c40d6529983c569e7004f32