Stable assigment of indices in the type order
Summary:
The previous code would assign indices sequentially in the type order;
since the insertion order is essentially random, this would cause the
final state to be unnpredictable.

Change insertion strategy to use the annotation hash to select the
type order index, with the intent of making the final representation
stable.

Reviewed By: sinancepel

Differential Revision: D13992721

fbshipit-source-id: 731e43796730b9515e7bbaf34637bb38dbab94f7