Rewrite missing global annotation logic to throw on Anys
Summary:
everythingisterrible

Thought throwing MissingGlobal errors on Anys would be a quick tweak, but the logic surrounding assignments and when to error turned out to be such a nightmare. :(

Needed to distinguish: when is the current assign missing an annotation vs. when is it a bare assign on something that has already been annotated elsewhere and doesn't need to be; which are globals, attributes or locals or type aliases, etc.,

Every change broke something so I mostly rewrote this logic to hopefully make more sense and clearly label what annotations are which and stop joining concepts like "target annotation if X condition else value annotation". Tried to leave some comments to explain

Reviewed By: dark

Differential Revision: D13970621

fbshipit-source-id: 155b6dc718fea1dc9dc4799537ae31dbdd84d11d