Allow annotations containing aliases to Any in strict
Summary:
This covers cases except for globals... that gets tricky because:

```
MyType = typing.Any

x: MyType  # Can detect that this is fine, since MyType is not Any
...
x = 1  # All I have now is the type that `x` resolves to, which is Any.
```

Because I don't have access to the original type expression at an arbitrary future location, I can't tell if I should throw an error on the assign. Either we give up error merging (and the knowledge that a global is say, a Union of int and string based on what it's been assigned to in various sources) or allowing aliased Anys isn't possible with our current framework. Suggestions welcome here

Reviewed By: dkgi

Differential Revision: D13983363

fbshipit-source-id: 811e2cd5395b7b3cca1063812db4d707350dc8f8