Treat literal attributes as explicitly annotated
Summary:
We just did this for globals.

We need to exclude dataclasses because the PEP explicitly says only annotated attributes are added to the init function as parameters; we also don't want to do this for anything nested in any branching logic to avoid things like:

```
def foo(self):
    if test:
        self.x = 1
    else:
        self.x = "hi"
```
Don't want to error that `x` has type `str` but is used as `int`.

Reviewed By: sinancepel

Differential Revision: D13992789

fbshipit-source-id: 073a4c164542b4ae7a01022efd9456a3b08a254f