Compute deferred requests lazily
Summary:
When receiving a request to purge several thousands files, the
previous code would take tens of minutes just to complete the
dependency scan. Since that portion of the update is quiet, the user
would wonder if pyre is working at all.

By computing deferred request lazily, the time needed to compute the
results is dramatically reduced, and the interactivity of the update
is preserved (since the client will print the status report).

Reviewed By: sinancepel

Differential Revision: D13967834

fbshipit-source-id: dc6263f992c6fcdee9e43510601aa27a59e0e832