unconditionally kill when no server was found in stop
Summary:
We would like `pyre stop` to work consistently and people to not have to
resort to `pyre kill` - ensuring that we run kill if `stop` was ran will make sure that we stop Pyre more consistently.

Of course, this will come at a cost of killing your pyre servers for other projects if we fail - however, I think it's a reasonable tradeoff for consistently stopping Pyre given that we frequently have cases where users have to run pyre kill after pyre stop fails.

Reviewed By: dark

Differential Revision: D14015742

fbshipit-source-id: 627a20fbb7c1092a92ff5b848815d9507ad8e09b