Return relative configuration path in upgrade
Summary:
Used to pass the entire path to the configuration file, which is not valid on different machines. We want the relative path instead.

Relative path was also being used improperly when generating the title of the commit - we want to relativize the full paths with each other to get the name.

Reviewed By: dkgi

Differential Revision: D14066922

fbshipit-source-id: d21d3fa85b6a1960d9b176bbfad2e383a5f92f10