Allow aliases to Any in global types
Summary:
Was done for the other missingX annotation errors here: D13983363

Globals are hard because we join their use cases to make coherent suggestions, and we'd prefer not to lose this feature:

```
MyType = typing.Any

x: MyType  # Can detect that this is fine, since MyType is not Any
...
x = 1  # All I have now is the type that `x` resolves to, which is Any.
```
(copied from previous diff)
Because I don't have access to the original type expression at an arbitrary future location, I can't tell if I should throw an error on the assign. Either we give up error merging (and the knowledge that a global is say, a Union of int and string based on what it's been assigned to in various sources) or allowing aliased Anys isn't possible with our current framework.

(solution here)
Keep track of whether we are throwing an error on the source of the global/attribute, since when it is defined we know whether or not we need to throw the error. We can still throw the helper errors elsewhere, but when we join them all together in `join_at_source`, we get one final error that we now know whether to display or not!

Reviewed By: dkgi

Differential Revision: D14035442

fbshipit-source-id: d8b8fad144116dab4ed6b354551ae59d16639ee0