Error on illegal annotation
Summary:
We used to indiscriminately add immutables for whatever target was given an explicit annotation.

Instead, we should error when annotations are given where they are not allowed, and skip adding these to globals. MyPy throws two distinct errors depending on if the bad target is an attribute or not, but I didn't think this was necessary so merged them into a single IllegalAnnotationTarget error.

Reviewed By: dkgi

Differential Revision: D14058242

fbshipit-source-id: dbdda4f839f1ba4dd39f9b550f64cde3fdf85270