Do not assume incomplete types from literals as explicit
Summary:
Started to see errors in distillery of the form:

```
x = None
...
x = 1
```
Erroring that global `x` (or attribute `x`) has type `none` but is used as type `int`.

Really, the assumed annotation from a literal assignment to None should be `Optional[Any]` since we'll be using this as an upper bound to compare future assignments to. But that's also not very useful to assume because it'll immediately throw an error saying you can't annotate this to contain Any (which is extra confusing given there is no visible annotation there).

This issue also occurs with other types including bottom that should be inferred as `Any`, such as `x = []`.

I think it's cleanest if we don't try to infer explicit annotations for types that contain Top, Any, or Bottom. The usual "missing global annotation" or "missing attribute annotation" errors will appear with the appropriate type suggestion instead.

Reviewed By: dkgi

Differential Revision: D14050883

fbshipit-source-id: 4786105bf9bf0b954b77d6064310dc2b531078cf