Use common merge function for merging Issues
Summary:
Since we no longer update old fields, the merge logic for issues is the same as everything else.

This does remove the warning when we have duplicate handles within the same run. The warning was added when we had issues with the same handle being produced due to a bug around the way we determine error codes from sources and sinks. Bug has been fixed.

Reviewed By: zlandau

Differential Revision: D14077827

fbshipit-source-id: d868a6eca84adcffa165f8185d2dded95a62b9cb