don't try to read configuration's that do not exist
Summary:
We were trying to read configurations which do not exist, causing `No configuration was found` messages to appear in sandcastle logs.

Check to make sure that the file exists before trying to read. This doesn't introduce TOCTTOU bugs because we still have the exception wrapping, it should be a usability improvement.

Now that the no configuration found message shouldn't appear in normal use, bump it to a warning.

Reviewed By: shannonzhu

Differential Revision: D14058606

fbshipit-source-id: b60afc58f3a45e6c61f557029024e30a76559382