don't try to read a local configuration at the top level
Summary:
When pyre is called, the first thing the python side does is changing directories to the directory with the .pyre_configuration. Reading .pyre_configuration.local will therefore always fail
UNLESS we have it in the same directory as the global configuration. In that case, I'd argue that you should just merge the two.

Reviewed By: dark

Differential Revision: D14078698

fbshipit-source-id: f0c85a45c9ab64a424ad9ab4fe418ccd87e20109