Pass traces and detail switches to error message generation
Summary:
Converting "show error traces" into "detailed" is confusing now, since error traces are a particular addition to the error but don't actually change the main error body (which we want to do with --hide-error-detail for upgrades).

Pass this flag into message generating, and rename some things to make the distinction clear.

Reviewed By: dkgi

Differential Revision: D14089101

fbshipit-source-id: 8a620899f00cc609f9c5d7ffb45282967dc353f4