Stop fusing fixmes
Summary:
Instead of trying to fuse fixmes:

```
line with error
```

If we have two errors on the same line we should do:

```
line with error
```

It's unlikely that a merged description will be short enough to fit on one line. Splitting just the description into multiple lines makes it less readable; now that we support fixmes skipping comment lines, we can keep the fixme and the code on the same line as its description.

Any fixmes with descriptions that are too long to be fused will still show on the same line:

```
line with three errors where 1 and 3 were both individually too long
```

Reviewed By: dkgi

Differential Revision: D14134572

fbshipit-source-id: 6277974fb2e8fb3502622ee78d247d0fec2e8a30