Replace potentially invalid generated assign with assert
Summary:
killme

Agonized over this one for a while -- the problem with the original code is:

```
def foo(e: Type[Exception]):
   try:
      ...
   except e as name:
      ...
```

gets preprocessed to a try body node linked to an except handling node with the body:

```
name: e = ...
....
```

This way we set the type of `name` to be whatever was given in the place of `e`.

The problem is, this is actually an invalid type since it's not registered as a type alias and we do correctly error on:

```
def foo(e: Type[int]):
   x: e = 1
```

Considered special-casing assignments where the value is ellipses, but that turns into a mess since these are used in stubs and elsewhere.
Considered turning the value of the assign into a special-case value like `$generated`, but that's also several locations of special casing that is ugly.
It's also hard to do an assign to `name = e()` instead, since we don't know what the initialization takes as parameters and we might have union types to handle, too.
The try block is also not available to actually match on in the backend, since we spit it up when creating the cfg and sticking the entire try back into the except branch is incorrect and ugly.
Realized in the end that assert is definitely the cleanest way to handle this, even if it's a little roundabout.

Reviewed By: dkgi

Differential Revision: D14151497

fbshipit-source-id: a2dddb7256a7b144572407881a65bee8d64eae39