Support multi-line pyre ignores
Summary:
Our fixme script typically outputs:

```
foo() # line with two errors
```

This should be acceptable, and not generate an unused ignore / unignored error, creating churn.

This diff makes ignores apply to the next line that is not a comment.

This will also allow:

```
foo()
```

Note: This may be useful for manual error suppression -- we are trying to move towards single line fixmes generated in the upgrade script to begin with.

Reviewed By: sinancepel

Differential Revision: D14114853

fbshipit-source-id: 4a2819806f80a258ec4c71459ed3ddfb8ae91390