Duplicate triage fields inside 'Issue'
Summary:
As part of merging Issue and IssueTriageInfo, the triage fields will eventually
be moved into Issue and IssueTriageInfo[Assoc] can go away. Meanwhile, during
migration, we need to run them side-by-side.

UI will write to both in the interim, then we do a one-time batch copy and get
rid of the old one.

We also need to make sure that table locking doesn't interfere with our writes
(will verify it when we are writing to both tables). I *think* it shouldn't be
a problem since we only lock rows that are being added, and we no longer update
old rows.

Reviewed By: zlandau

Differential Revision: D14095586

fbshipit-source-id: 1e7209676ed8cf8b728528612fafc3dbce4e8667