Move solve constraints into typeorder, and make constructor lookup available to all other functions
Summary: Ultimately the typeorder is incomplete without constructor lookup, as the relationship in the order between `Type[T]` and a callable is determined by the signatures of its constructors.  Thus we need access to it in order to ultimately make leq join and meet correct.  This was also a convenient time to move solve in, as we will need solve in order to correctly implement leq between generic and concrete callables.

Reviewed By: dkgi

Differential Revision: D13926954

fbshipit-source-id: 59014828454cba42686ea5e31b591ae326b97e80