Extend special casing of Dict[str,Any] to mapping
Summary:
We encourage users to use Mapping to avoid variance issues; in order to support JSON and other str->Any mappings we're special casing to allow Dict[str, Any] in strict. Really this should be allowing any subtype of Mapping[str, Any].

Had to make some server test changes because `typing.Mapping` was not in the typeorder in most of the tests. Added basic test.ml stubs for typing/builtins to avoid these crashes.

Reviewed By: dkgi

Differential Revision: D14195016

fbshipit-source-id: 789df455eb49897df168e082a7be9365044f074f