fix namespace collisions for stubs under local root
Summary:
I initially thought this was a problem with pyi files getting parsed in the local root due to the extensions feature, but that wasn't the case. The issue was when we had a directory in the search path be under the local root, in which case we also collected that stub when collecting stubs from the Pyre root. This led to inconsistencies when you had interfering stubs within the search path/local root - since we prioritize local root stubs, if we had interfering stubs under the local root, we'd pick whichever one happened to get returned first from Path.list.

Make sure this doesn't happen by relying on handles (which are the canonical way of determining which qualifier a file has based on our search path) and only adding a handle if the reconstructed version (root/handle) matches the original path.

Reviewed By: dkgi

Differential Revision: D14212022

fbshipit-source-id: eb1c693a1907e020a00deb2417be0d9f50f60226