Display number of branches at a trace frame
Summary:
For example, if edges A -> B and A -> C exist, then display the number 2 beside the trace frames for B and C. Only display this number if it is greater than 1.

This is the first step in navigating branched traces. The idea is that if the active trace frame is one with multiple branches, then the user can run a command to show the other branch options and choose which one to go to.

Reviewed By: zlandau

Differential Revision: D14210509

fbshipit-source-id: ce0765d8baa736f4f0dfd2390e479cdcc0b997d4