Relaxing standards for typeddicts in preprocessing
Summary: Its a bad experience to have a declaration issue with your typeddict and not have any messaging about it and have downstream errors.  Earlier I thought that this was the only way to ever get the errors to a future error handler, but I realized now we can transform it and leave the original in, but marked with something (a decorator, a qualification on the name) that makes it get checked downstream.

Reviewed By: shannonzhu

Differential Revision: D14219692

fbshipit-source-id: c6aa595c9ea8c65f034687b13a05d3f785dae357