making all joins true joins
Summary:
There were a whole variety of issues being suppressed by this behavior, especially with supression of mismatch with any errors.  There are obviously some presentation issues now with showing mismatches of long parametric types without any guidance as to where to look but that's a separate issue.

This was also screwing stuff up with literals, as we do want to enforce a more relaxed join discipline, i.e. join 1 7 giving int not union[1, 7] but the true join was bringing it back.  putting the special case there as well seemed ridiculous.

Reviewed By: shannonzhu

Differential Revision: D14185551

fbshipit-source-id: 6cda05d1e261c4a14c54ffcd4cca3f52d6fa7757