parse `scipy/stats/_continuous_distns.py`
Summary:
We were failing to parse `scipy/stats/_continuous_distns.py`, as it is a UTF-8 file with a byte order mark.

Byte order marks are not recommended for UTF-8, but windows does it anyway: https://en.wikipedia.org/wiki/Byte_order_mark#UTF-8

As for the question of "why don't you rely on Uchar, the unicode library for ocaml?" - the library doesn't come with a way of serializing unicode characters to ascii, so I'd have to implement that mechanism myself anyway. This seemed like a cleaner solution compared to implementing the "divide/mod by 256 until you get 0 to get the byte representation" solution.

Reviewed By: shannonzhu

Differential Revision: D14265939

fbshipit-source-id: c1f5e4b4393a5acdf1be697e1f26ffa694396fe9